requirements.txt
asgiref==3.2.7
colorama==0.4.6
distutils-pytest==0.2.1
Django==3.0.5
django-model-utils==4.0.0
djangorestframework==3.11.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
pytz==2019.3
setuptools==68.2.2
sqlparse==0.3.1
Unipath==1.1
